requirements.txt
web3==6.11.1
requests==2.31.0
python-dotenv==1.0.0
eth-account==0.10.0
cryptography==41.0.7 